// src/types.ts
export interface ISpecs {
  SW_m: number;
  SH_m: number;
  SW_ft: number;
  SH_ft: number;
  SW_in: number;
  SH_in: number;
  SA_m: number;
  SA_ft: number;
  DDS_m: number;
  DDS_in: number;
  AspectRatio: number[];
  CH: number;
  CV: number;
  TC: number;
  SHR: number;
  SVR: number;
  STR: number;
}